docs(experiments): make basic example use nested spans
